+++ examples/ls.zig
@@ -14,7 +14,7 @@ pub fn main() anyerror!void {
 
     var update_cmd = app.createCommand("update", "Update the app or check for new updates");
     try update_cmd.addArg(Arg.booleanOption("check-only", null, "Only check for new update"));
-    try update_cmd.addArg(Arg.singleArgumentOptionWithValidValues("branch", 'b', "Branch to update", &[_][]const u8{
+    try update_cmd.addArg(Arg.singleValueOptionWithValidValues("branch", 'b', "Branch to update", &[_][]const u8{
         "stable",
         "nightly",
         "beta",
@@ -29,7 +29,7 @@ pub fn main() anyerror!void {
     try myls.addArg(Arg.booleanOption("version", null, null));
     try myls.addArg(Arg.singleValueOption("ignore", 'I', null));
     try myls.addArg(Arg.singleValueOption("hide", null, null));
-    try myls.addArg(Arg.singleArgumentOptionWithValidValues("color", 'C', null, &[_][]const u8{
+    try myls.addArg(Arg.singleValueOptionWithValidValues("color", 'C', null, &[_][]const u8{
         "always",
         "auto",
         "never",